+++ src/pymyhondaplus/__init__.py
@@ -3,6 +3,7 @@
 from .api import AuthTokens, CommandResult, EVStatus, Geofence, HondaAPI, HondaAPIError, HondaAuthError, SubscriptionService, Subscription, UIConfiguration, UserProfile, Vehicle, VehicleCapabilities, compute_trip_stats, parse_charge_schedule, parse_climate_schedule, parse_ev_status
 from .auth import DeviceKey, HondaAuth, encrypt_request
 from .storage import SecretStorage, get_storage
+from .translations import TRANSLATIONS, get_translator
 
 __all__ = [
     "AuthTokens",
@@ -11,6 +12,7 @@ __all__ = [
     "Geofence",
     "Subscription",
     "SubscriptionService",
+    "TRANSLATIONS",
     "UIConfiguration",
     "UserProfile",
     "Vehicle",
@@ -19,6 +21,7 @@ __all__ = [
     "HondaAPIError",
     "HondaAuthError",
     "compute_trip_stats",
+    "get_translator",
     "parse_charge_schedule",
     "parse_climate_schedule",
     "parse_ev_status",